If $COMPOSER_BIN_DIR is defined, remove it from the beginning of $PATH
When running Composer scripts, Composer will prepend its bin-dir to the user's `$PATH`.

However, this causes issues when we're running a script named "shellcheck" and we're looking
for "shellcheck" in the user's path.

If we find the `$COMPOSER_BIN_DIR` environment variable, strip it from the beginning of the
user's `$PATH` so we can find the *real* ShellCheck.

Fixes #2.