Look for COMPOSER_RUNTIME_BIN_DIR, not COMPOSER_BIN_DIR
[Since Composer 2.2.6](https://getcomposer.org/doc/articles/vendor-binaries.md#finding-the-composer-bin-dir-from-a-binary) (the latest version), Composer exposes the `COMPOSER_RUNTIME_BIN_DIR` environment variable when running as a vendor script.

However, there are a few conditions under which this variable may not be available:

* When running binaries defined by the root package
* When running on Composer < 2.2.6
* When executing the script outside of a Composer script context

This commit replaces the check for `COMPOSER_BIN_DIR` with `COMPOSER_RUNTIME_BIN_DIR` and adds a fallback that uses `composer config --absolute bin-dir`; if we're able to determine a project's Composer bin-dir, we'll strip this off the front of a user's `$PATH`.

(Re-)fixes #2.